@@ -1,4 +1,7 @@
 import 'package:flutter/material.dart';
+import 'quiz_brain.dart';
+
+QuizBrain quizBrain = QuizBrain();
 
 void main() => runApp(Quizzler());
 
@@ -25,6 +28,46 @@ class QuizPage extends StatefulWidget {
 }
 
 class _QuizPageState extends State<QuizPage> {
+
+    List<Icon> scoreKeeper = [
+    // Icon(            
+    //   Icons.check,
+    //   color: Colors.green,
+    //   ),
+    // Icon(
+    //   Icons.close,
+    //   color: Colors.red,
+    // )      
+  ];
+
+  void checkAnswer(bool userSelectedAnswer){
+    bool correctAnswer = quizBrain.getAnswer();
+
+    if (correctAnswer == userSelectedAnswer){
+      print("User got it right."); 
+      scoreKeeper.add(
+        Icon(            
+          Icons.check,
+          color: Colors.green,
+          ),
+      );                 
+    } 
+    else{
+      print("User got it wrong.");
+      scoreKeeper.add(
+        Icon(            
+          Icons.close,
+          color: Colors.red,
+          ),
+      ); 
+    }
+
+    setState(() {
+      quizBrain.nextQuestion();
+    });
+  }
+   
+
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -37,7 +80,7 @@ class _QuizPageState extends State<QuizPage> {
             padding: EdgeInsets.all(10.0),
             child: Center(
               child: Text(
-                'This is where the question text will go.',
+                quizBrain.getQuestionText(),
                 textAlign: TextAlign.center,
                 style: TextStyle(
                   fontSize: 25.0,
@@ -50,41 +93,47 @@ class _QuizPageState extends State<QuizPage> {
         Expanded(
           child: Padding(
             padding: EdgeInsets.all(15.0),
-            child: FlatButton(
-              textColor: Colors.white,
-              color: Colors.green,
-              child: Text(
-                'True',
-                style: TextStyle(
-                  color: Colors.white,
-                  fontSize: 20.0,
+            child: Container(              
+              color: Colors.green,    
+              child: TextButton(
+                //textColor: Colors.white,
+                child: Text(
+                  'True',
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 20.0,
+                  ),
                 ),
+                onPressed: () {
+                  //The user picked true.
+                  checkAnswer(true);
+                },
               ),
-              onPressed: () {
-                //The user picked true.
-              },
             ),
           ),
         ),
         Expanded(
           child: Padding(
             padding: EdgeInsets.all(15.0),
-            child: FlatButton(
+            child: Container(
               color: Colors.red,
-              child: Text(
-                'False',
-                style: TextStyle(
-                  fontSize: 20.0,
-                  color: Colors.white,
+              child: TextButton(
+                child: Text(
+                  'False',
+                  style: TextStyle(
+                    fontSize: 20.0,
+                    color: Colors.white,
+                  ),
                 ),
+                onPressed: () {
+                  //The user picked false.
+                  checkAnswer(false);
+                },
               ),
-              onPressed: () {
-                //The user picked false.
-              },
             ),
           ),
         ),
-        //TODO: Add a Row here as your score keeper
+        Row(children: scoreKeeper,)
       ],
     );
   }
